Refactor data structures into thier own module
- preparation for a more view based data access
